chore: protect routes

+++ src/pages/api/routes/programs/[id].ts
@@ -2,6 +2,9 @@ import type { NextApiRequest, NextApiResponse } from "next";
 
 import { prisma } from "../../../../auth/db";
 
+import { getServerSession } from "next-auth";
+import { authOptions } from 'auth/auth';
+
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   const id = req.query.id as string;
 
@@ -37,6 +40,11 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   }
 
   if (req.method === "DELETE") {
+    const session = await getServerSession(req, res, authOptions);
+    if (!session) {
+      res.status(401).json({ message: `Unauthorized.` });
+      return;
+    }
     try {
       const program = await prisma.onboardingProgram.delete({
         where: {
@@ -56,6 +64,11 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   }
 
   if (req.method === "PATCH") {
+    const session = await getServerSession(req, res, authOptions);
+    if (!session) {
+      res.status(401).json({ message: `Unauthorized.` });
+      return;
+    }
     try {
       const { name, content } = req.body as {
         name: string;
